Updates from Ian Curtis (i.curtis on gmail): - ported the project to visual studio 2008 - fixed a tonne of compile warnings, mostly related to depreciated API stuff - fixed 1 compile error
git-svn-id: svn+ssh://svn.code.sf.net/p/winlirc/code/trunk@62 92c2518b-485e-4ca2-9c37-138eef3f838b

+++ winlirc/irdriver.cpp
@@ -275,7 +275,7 @@ void CIRDriver::ThreadProc(void)
 			prev=dcd;
 		}
 
-		int deltv=lr_time-lr_lasttime;
+		int deltv=(int)(lr_time-lr_lasttime);
 		if (devicetype && (deltv>15)) {		
 			/* More than 15 seconds passed */
 			deltv=0xFFFFFF;
